+++ README.md
@@ -1,7 +1,7 @@
-1. Install all node dependencies in submodules (yes its a bit tedious)
-3. Change port/public ip in index.js if needed
-4. Unzip encrypted "secrets.zip" (use zip linux package -> will end up with "secrets" folder in main dir)
-5. Run index.js!
+1. Install all node dependencies
+2. Change port/public ip in index.js if needed
+3. Unzip encrypted "secrets.zip" (use zip linux package -> will end up with "secrets" folder in main dir)
+4. Run index.js!
 
 Sites: /grid /dash /dashboard (old, forgotten, use /dash instead)
 API: /loggedin /clock
